Fix bug
Users tring to update their email could not do it, since my updateUser
function was retrieving users by email and this new email was (normally)
not already in the database, therefore I would get "no user" error.
Therefore updated to finding users by Id, which are encoded in the
token.